requirements.txt
autopep8==2.0.0
Pillow==9.3.0
pycodestyle==2.9.1
